# packages/platform/lib/openapi-spec.yaml
openapi: 3.0.3
info:
  title: ElizaOS Platform API
  description: |
    The ElizaOS Platform API provides programmatic access to manage AI agents, organizations, API keys, and billing.

    ## Authentication

    The API supports two authentication methods:

    1. **Bearer Token (JWT)**: Include the access token in the Authorization header
       ```
       Authorization: Bearer <access_token>
       ```

    2. **API Key**: Include your API key in the X-API-Key header
       ```
       X-API-Key: pk_live_...
       ```
  version: 1.0.0
  contact:
    email: support@elizaos.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3333/api/v1
    description: Development server
  - url: https://api.elizaos.ai/v1
    description: Production server
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
  /auth/login:
    get:
      tags:
        - Authentication
      summary: Initiate SSO login
      description: Get the WorkOS SSO authorization URL to start the login flow
      operationId: initiateLogin
      security: []
      parameters:
        - name: organization_id
          in: query
          description: Organization ID for SSO (optional)
          schema:
            type: string
        - name: redirect_uri
          in: query
          description: Custom redirect URI after authentication
          schema:
            type: string
            format: uri
      responses:
        '200':
          description: Login URL generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Get the authenticated user's profile and organization
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login or refresh endpoints

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key starting with 'pk_'
...
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          $ref: '#/components/schemas/Error'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        avatar:
          type: string
          format: uri
        emailVerified:
          type: boolean
        lastLoginAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    Organization:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        domain:
          type: string
        subscriptionTier:
          type: string
          enum: [free, pro, premium, enterprise]
        maxAgents:
          type: integer
        maxUsers:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                authUrl:
                  type: string
                  format: uri
                state:
                  type: string
...
    CurrentUserResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                user:
                  $ref: '#/components/schemas/User'
                organization:
                  $ref: '#/components/schemas/Organization'
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error:
              code: AUTHENTICATION_REQUIRED
              message: Authentication is required
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error:
              code: INTERNAL_SERVER_ERROR
              message: An unexpected error occurred